requirements.txt
# ── Kirana Store — Root Python Requirements ─────────────────────────────────
# Install per-service: cd backend/<service> && pip install -r requirements.txt
# This file lists ALL backend deps combined for reference / CI

# ── Core Framework ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── ORM / Database ───────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Request Parsing ──────────────────────────────────────────────────────────
python-multipart>=0.0.9

# ── Dev / Testing ────────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0
